pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.46.0",
    "transformers>=4.52.4",
]
